Split the ATT implementation into multiple files
